ajuste no desafio6

+++ desafio6.py
@@ -1,4 +1,4 @@
-numero = int(input('Insira um número: '))
+numero = float(input('Insira um número: '))
 while numero != 10:
-    numero = int(input('Insira um número: '))
+    numero = float(input('Insira um número: '))
 print ("Número 10!")
